ratings working, sending data to DB working
