+++ src/components/LogIns/ProfileInformation.js
@@ -1,4 +1,3 @@
-/* eslint-disable react/button-has-type */
 import React, { useEffect, useContext, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { CornerDownLeft } from 'react-feather';
